refactor recover and rollup exexutor test
