+++ src/node/src/lib.rs
@@ -19,6 +19,7 @@ pub mod ar_toolbox;
 pub mod command;
 pub mod indexer_impl;
 mod mutation_utils;
+mod node_test_base;
 pub mod recover;
 pub mod rollup_executor;
 pub mod storage_node_light_impl;
