+++ src/node/src/rollup_executor.rs
@@ -27,7 +27,10 @@ use db3_storage::system_store::{SystemRole, SystemStore};
 use std::sync::atomic::{AtomicU64, Ordering};
 use std::sync::Arc;
 use std::time::Instant;
-use tracing::{info, warn};
+#[cfg(test)]
+use std::{println as info, println as warn};
+#[cfg(not(test))]
+use tracing::{info, warn}; // Workaround to use prinltn! for logs.
 
 #[derive(Clone)]
 pub struct RollupExecutorConfig {
@@ -268,7 +271,7 @@ impl RollupExecutor {
                 .storage
                 .get_range_mutations(last_end_block, current_block)?;
             if mutations.len() <= 0 {
-                info!("no block to rollup");
+                info!("no mutations to rollup");
                 return Ok(());
             }
             self.pending_mutations
@@ -335,134 +338,22 @@ impl RollupExecutor {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use db3_crypto::db3_address::DB3Address;
-    use db3_proto::db3_base_proto::SystemConfig;
-    use db3_proto::db3_mutation_v2_proto::MutationAction;
-    use db3_storage::db_store_v2::{DBStoreV2, DBStoreV2Config};
-    use db3_storage::doc_store::DocStoreConfig;
-    use db3_storage::mutation_store::MutationStoreConfig;
-    use db3_storage::state_store::StateStore;
-    use db3_storage::state_store::StateStoreConfig;
-    use db3_storage::system_store::SystemStoreConfig;
-    use std::path::PathBuf;
+    use crate::node_test_base::NodeTestBase;
+    use crate::rollup_executor;
     use tempdir::TempDir;
-    fn generate_config(
-        real_path: &str,
-    ) -> (
-        StateStoreConfig,
-        SystemStoreConfig,
-        MutationStoreConfig,
-        RollupExecutorConfig,
-        DBStoreV2Config,
-    ) {
-        if let Err(_e) = std::fs::create_dir_all(real_path) {}
-        let path = PathBuf::from(env!("CARGO_MANIFEST_DIR"));
-        let key_root_path = path
-            .parent()
-            .unwrap()
-            .parent()
-            .unwrap()
-            .join("tools/keys")
-            .to_str()
-            .unwrap()
-            .to_string();
-        let rollup_config = RollupExecutorConfig {
-            temp_data_path: format!("{real_path}/data_path"),
-            key_root_path: key_root_path.to_string(),
-        };
-        let system_store_config = SystemStoreConfig {
-            key_root_path: key_root_path.to_string(),
-            evm_wallet_key: "evm".to_string(),
-            ar_wallet_key: "ar".to_string(),
-        };
-
-        let store_config = MutationStoreConfig {
-            db_path: format!("{real_path}/mutation_path"),
-            block_store_cf_name: "block_store_cf".to_string(),
-            tx_store_cf_name: "tx_store_cf".to_string(),
-            rollup_store_cf_name: "rollup_store_cf".to_string(),
-            gc_cf_name: "gc_store_cf".to_string(),
-            message_max_buffer: 4 * 1024,
-            scan_max_limit: 50,
-            block_state_cf_name: "block_state_cf".to_string(),
-        };
-        let state_config = StateStoreConfig {
-            db_path: format!("{real_path}/state_store"),
-        };
-
-        let db_store_config = DBStoreV2Config {
-            db_path: format!("{real_path}/db_store"),
-            db_store_cf_name: "db".to_string(),
-            doc_store_cf_name: "doc".to_string(),
-            collection_store_cf_name: "cf2".to_string(),
-            index_store_cf_name: "index".to_string(),
-            doc_owner_store_cf_name: "doc_owner".to_string(),
-            db_owner_store_cf_name: "db_owner".to_string(),
-            scan_max_limit: 50,
-            enable_doc_store: false,
-            doc_store_conf: DocStoreConfig::default(),
-            doc_start_id: 1000,
-        };
-
-        (
-            state_config,
-            system_store_config,
-            store_config,
-            rollup_config,
-            db_store_config,
-        )
-    }
-
-    async fn setup_for_smoke_test() -> Result<RollupExecutor> {
-        let tmp_dir_path = TempDir::new("add_store_path").expect("create temp dir");
-        let real_path = tmp_dir_path.path().to_str().unwrap().to_string();
-        let (state_config, system_store_config, store_config, rollup_config, _) =
-            generate_config(real_path.as_str());
-        let state_store = Arc::new(StateStore::new(state_config).unwrap());
-        let system_store = Arc::new(SystemStore::new(system_store_config, state_store.clone()));
-        let storage = MutationStore::new(store_config).unwrap();
-        storage.recover().unwrap();
-        let system_config = SystemConfig {
-            min_rollup_size: 1024,
-            rollup_interval: 1000,
-            network_id: 1,
-            evm_node_url: "ws://127.0.0.1:8545".to_string(),
-            ar_node_url: "http://127.0.0.1:1985".to_string(),
-            chain_id: 31337_u32,
-            rollup_max_interval: 2000,
-            contract_addr: "0x5FbDB2315678afecb367f032d93F642f64180aa3".to_string(),
-            min_gc_offset: 100,
-        };
-        let result = system_store.update_config(&SystemRole::DataRollupNode, &system_config);
-        assert_eq!(true, result.is_ok());
-        for _i in 0..1000 {
-            let payload: Vec<u8> = vec![1];
-            let signature: &str = "0xasdasdsad";
-            let (_id, block, order) = storage
-                .generate_mutation_block_and_order(payload.as_ref(), signature)
-                .unwrap();
-            let result = storage.add_mutation(
-                payload.as_ref(),
-                signature,
-                "",
-                &DB3Address::ZERO,
-                1,
-                block,
-                order,
-                1,
-                MutationAction::CreateDocumentDb,
-            );
-            assert_eq!(true, result.is_ok());
-        }
-        RollupExecutor::new(rollup_config, storage, system_store).await
-    }
 
     #[tokio::test]
     async fn test_rollup_smoke_test() {
-        match setup_for_smoke_test().await {
-            Ok(rollup_executor) => {
+        let tmp_dir_path = TempDir::new("test_rollup_smoke_test").expect("create temp dir");
+        match NodeTestBase::setup_for_smoke_test(&tmp_dir_path).await {
+            Ok((rollup_executor, recover)) => {
                 let result = rollup_executor.process().await;
                 assert_eq!(true, result.is_ok());
+                let result = recover.get_latest_arweave_tx().await;
+                assert_eq!(true, result.is_ok());
+                let tx = result.unwrap();
+                println!("the tx is {}", tx);
+                assert!(!tx.is_empty());
             }
             Err(e) => {
                 assert!(false, "{e}");
